metrics: 1. 为0的number不输出 2. 将value进行generics处理，简化输出； 3. 简化metric_name的处理，减少一次copy

+++ metrics/src/types/rtt.rs
@@ -33,7 +33,7 @@ impl Rtt {
             }
             let max = self.max.zero();
             if max > 0 {
-                w.write(&id.path, id.key, "max_us", max as f64);
+                w.write(&id.path, id.key, "max_us", max);
             }
         }
     }
